PIL-1804: Added link from submission history page to due and overdue returns page (#44)
We need to add the URL into the link placeholder at the bottom of the
submission history page (/submission-history) in the submission
frontend. The submission history page and the due and overdue returns
page were developed in tandem so this was unable to be implemented at
the time. This should be a minimal change from the dev perspective but
this will also impact the test automation scenarios.

This should link to the due and overdue returns page:
/due-and-overdue-returns