PIL-1721: Content Update for BTN start page (#49)
1. The page should be accessible via
url:/below-threshold-notification/start
2. Additional information clarifies that an information return (IR) is
not needed when BTN is submitted"HMRC will not expect to receive an
information return while your group is below- threshold"
3. Include content to H2: "If you need to submit a UK tax return for
this accounting period, you do not qualify for a Below-Threshold
Notification.
4. Remove the 3&4 eligibility criteria under H3
5. The last paragraph should include "If you submit a Below-Threshold
Notification, this will replace any returns you've submitted for that
period. It will also replace any returns you have already submitted for
your most recent account period."
6. The page should display ALL the current content, including the
various sections, such as H1, H2 and H3, as shown in the TO-BE prototype
7. The"Back" button should allow the user to return to the previous
step/Page.
8. The "Green Continue" button should be functional for users to
navigate to the next page