PIL-1913: Below-Threshold Notification (BTN) start page-(Agents View) (#52)
Build the Agent's view start page for the BTN Journey 
Please note: toggled off for feature release

The page should be accessible via
url:/below-threshold-notification/start
The page should display ALL the content including the various sections
such as H1, H2 and H3, with eligibility sections as shown in the
prototype.
The"Back" button should allow the user to return to the previous
step/Page.
The "Green Continue" button should be functional for users to navigate
to the next page
 
Prototype:
https://pillar2-prototype.herokuapp.com/current/below-threshold-notification/submit/agent/start

---------

Co-authored-by: MoDweik <180322631+MoDweik@users.noreply.github.com>