PIL-1915: Build BTN screens For both single and multiple Accounting Period-Agents View (#60)
For groups that does not meet the revenue threshold for pillar 2 top-up
taxes in a given account period, the authorised agent will submit a
Below-threshold notification (BTN) for an account period. This ticket is
to build the account period pages for the BTN Journey.

Build new pages in line with prototype
Amend content for existing pages

Scenarios in ticket:
https://jira.tools.tax.service.gov.uk/browse/PIL-1915